requirements.txt
customtkinter
pytest
pillow
